Skip build in case the library is already installed.

+++ scripts/macosx-build-dependencies.sh
@@ -46,6 +46,12 @@ printUsage()
 build_qt()
 {
   version=$1
+
+  if [ -d $DEPLOYDIR/lib/QtCore.framework ]; then
+    echo "qt already installed. not building"
+    return
+  fi
+
   echo "Building Qt" $version "..."
   cd $BASEDIR/src
   rm -rf qt-everywhere-opensource-src-$version
@@ -75,6 +81,12 @@ build_qt()
 build_gmp()
 {
   version=$1
+
+  if [ -f $DEPLOYDIR/lib/libgmp.dylib ]; then
+    echo "gmp already installed. not building"
+    return
+  fi
+
   echo "Building gmp" $version "..."
   cd $BASEDIR/src
   rm -rf gmp-$version
@@ -171,6 +183,12 @@ EOF
 build_mpfr()
 {
   version=$1
+
+  if [ -f $DEPLOYDIR/include/mpfr.h ]; then
+    echo "mpfr already installed. not building"
+    return
+  fi
+
   echo "Building mpfr" $version "..."
   cd $BASEDIR/src
   rm -rf mpfr-$version
@@ -211,6 +229,12 @@ build_mpfr()
 build_boost()
 {
   version=$1
+
+  if [ -f $DEPLOYDIR/lib/libboost_system.dylib ]; then
+    echo "boost already installed. not building"
+    return
+  fi
+
   bversion=`echo $version | tr "." "_"`
   echo "Building boost" $version "..."
   cd $BASEDIR/src
@@ -248,6 +272,12 @@ build_boost()
 build_cgal()
 {
   version=$1
+
+  if [ -f $DEPLOYDIR/lib/libCGAL.dylib ]; then
+    echo "cgal already installed. not building"
+    return
+  fi
+
   echo "Building CGAL" $version "..."
   cd $BASEDIR/src
   rm -rf CGAL-$version
@@ -279,6 +309,12 @@ build_cgal()
 build_glew()
 {
   version=$1
+
+  if [ -f $DEPLOYDIR/lib/libGLEW.dylib ]; then
+    echo "glew already installed. not building"
+    return
+  fi
+
   echo "Building GLEW" $version "..."
   cd $BASEDIR/src
   rm -rf glew-$version
@@ -358,6 +394,12 @@ build_glib2()
 build_opencsg()
 {
   version=$1
+
+  if [ -f $DEPLOYDIR/lib/libopencsg.dylib ]; then
+    echo "opencsg already installed. not building"
+    return
+  fi
+
   echo "Building OpenCSG" $version "..."
   cd $BASEDIR/src
   rm -rf OpenCSG-$version
@@ -377,6 +419,16 @@ build_opencsg()
 build_eigen()
 {
   version=$1
+
+  if [ -d $DEPLOYDIR/include/eigen2 ]; then
+    echo "eigen2 already installed. not building"
+    return
+  fi
+  if [ -d $DEPLOYDIR/include/eigen3 ]; then
+    echo "eigen3 already installed. not building"
+    return
+  fi
+
   echo "Building eigen" $version "..."
   cd $BASEDIR/src
   rm -rf eigen-$version
